+++ RELEASE_NOTES.md
@@ -1,4 +1,10 @@
 ### 22.1.2016
+### equitania, 1.0.163
+#### FIX
+-changes for product.template,product.product,product.template.standard_price_history,
+product.pricelist.item,eq_pricelist_item_search and product.category object
+
+### 22.1.2016
 ### equitania, 1.0.162
 #### FIX
 -changes for sale.order and sale.order.line object